+++ Examples/NormalizationRatioTests.m
@@ -19,13 +19,13 @@ fprintf('**************************************************************\n');
 % Initialize the analytical solution
 n = 2*64;
 latitude = 33;
-[rhoFunction, N2Function, zIn] = InternalModes.StratificationProfileWithName(profiles{iProfile});
-z = linspace(min(zIn),max(zIn),n)';
-if strcmp(profiles{iProfile},'constant')==1
-    imAnalytical = InternalModesConstantStratification(5.2e-3,zIn,z,latitude,'nModes',n);
-else
-    imAnalytical = InternalModesExponentialStratification([5.2e-3 1300],zIn,z,latitude,'nModes',n);
-end
+    [rhoFunction, N2Function, zIn] = InternalModes.StratificationProfileWithName(profiles{iProfile});
+    z = linspace(min(zIn),max(zIn),n)';
+    if strcmp(profiles{iProfile},'constant')==1
+        imAnalytical = InternalModesConstantStratification(N0=5.2e-3,zIn=zIn,zOut=z,latitude=latitude,nModes=n);
+    else
+        imAnalytical = InternalModesExponentialStratification(N0=5.2e-3,b=1300,zIn=zIn,zOut=z,latitude=latitude,nModes=n);
+    end
 imAnalytical.upperBoundary = upperBoundary;
 imAnalytical.normalization = Normalization.kConstant;