+++ Extras/SQGModeTest.m
@@ -10,17 +10,17 @@ z = linspace(zIn(1),0,n)';
 % rho = @(z) -(N0*N0*rho_0/g)*z + rho_0;
 % N2 = @(z) N0*N0*ones(size(z));
 % 
-% im = InternalModesConstantStratification( rho, zIn, z, lat );
+% im = InternalModesConstantStratification(N0=N0,rho0=rho_0,zIn=zIn,zOut=z,latitude=lat);
 % 
 k = 10.^linspace(log10(1e-5),log10(1e-1),10)';
 % psi_t = im.SurfaceModesAtWavenumber( k );
 % psi_b = im.BottomModesAtWavenumber( k );
 % 
-% im_spec = InternalModesFiniteDifference( rho, zIn, z, lat );
+% im_spec = InternalModesFiniteDifference(rho=rho,zIn=zIn,zOut=z,latitude=lat);
 % psi_t_spec = im_spec.SurfaceModesAtWavenumber( k );
 % psi_b_spec = im_spec.BottomModesAtWavenumber( k );
 % 
-% im_wkb = InternalModesWKB( rho, zIn, z, lat );
+% im_wkb = InternalModesWKB(rho=rho,zIn=zIn,zOut=z,latitude=lat);
 % psi_t_wkb = im_wkb.SurfaceModesAtWavenumber(k(1));
 % 
 % figure
@@ -38,8 +38,8 @@ rho = @(z) rho_0*(1 + L_gm*N0*N0/(2*g)*(1 - exp(2*z/L_gm)));
 N2Func = @(z) N0*N0*exp(2*z/L_gm);
 zIn = [-5000 0];
 
-im_spec = InternalModesSpectral( rho, zIn, z, lat );
-im_wkb = InternalModesWKB( rho, zIn, z, lat );
+im_spec = InternalModesSpectral(rho=rho,zIn=zIn,zOut=z,latitude=lat);
+im_wkb = InternalModesWKB(rho=rho,zIn=zIn,zOut=z,latitude=lat);
 
 ik = 1;
 alpha = 2/L_gm;