+++ Extras/WKBModeTest.m
@@ -7,12 +7,12 @@ N0 = 5.2e-3;
 zOut = linspace(min(zIn),max(zIn),500)';
 latitude = 33;
 
-im = InternalModesWKB(rhoFunction, zIn, zOut, latitude);
+im = InternalModesWKB(rho=rhoFunction,zIn=zIn,zOut=zOut,latitude=latitude);
 omega = im.f0 + 0.4*(N0-im.f0);
 [F, G, h] = im.ModesAtFrequency(omega);
 
 
-imAnalytical = InternalModesWKBSpectral(rhoFunction, zIn, zOut, latitude,'nEVP',512);
+imAnalytical = InternalModesWKBSpectral(rho=rhoFunction,zIn=zIn,zOut=zOut,latitude=latitude,nEVP=512);
 [F_analytical, G_analytical, h_analytical] = imAnalytical.ModesAtFrequency(omega);
 
 iMode = 4;